refactor, dev: вынес некоторый код из src.utils в src.business_logic.utils
